==> notebooks_clean/02_EDA_categorical.ipynb <==
import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns
from avp_pckg.DataFrame import AvPdataFrame 

%load_ext autoreload
%autoreload 2


df = pd.read_csv('data\\df_train.csv', parse_dates=['PurchDate'], index_col=0)
# display(df.head())
# df.info()

# skiped:'Auction',  'VehYear', 'Transmission', 'WheelTypeID', 'Nationality', 'TopThreeAmericanName', 'IsOnlineSale', 'VNST',

cols = ['IsBadBuy', 'VehicleAge', 'Make',
       'Model', 'Trim', 'SubModel', 'Color',
       'WheelType',  'Size',
       'PRIMEUNIT', 'AUCGUART', 'BYRNO', 'VNZIP1',
        'Auction',  'VehYear', 'Transmission', 'WheelTypeID', 'Nationality', 'TopThreeAmericanName', 'IsOnlineSale', 'VNST',
        ]
target_col = 'IsBadBuy'
df = df.loc[:, cols]
df.fillna('empty', inplace=True)
df.info()

df_noWheels = AvPdataFrame(df[df['WheelType'] == 'empty'].copy()) 
df_Wheels = AvPdataFrame(df[df['WheelType'] != 'empty'].copy())

df_noWheels.columns

cols = ['VehicleAge', 'Make',
       'Model', 'Trim', 'SubModel', 'Color',
        'Size',
       'PRIMEUNIT', 'AUCGUART', 'BYRNO', 'VNZIP1', 
       'Auction', 'VehYear', 'Transmission', 'Nationality', 'TopThreeAmericanName', 'IsOnlineSale', 'VNST',
        ]
# df_noWheels.info()
df_noWheels.cols_importance(cols=cols, target='IsBadBuy').sort_values('delta100', ascending=False)

df_noWheels.calc_frequency(col='Auction', target=target_col)

### Conclusion:
- 'Auction' column very important for no-Wheels data! 

df_noWheels.calc_frequency(col='AUCGUART', target=target_col)
df_noWheels.calc_frequency(col='PRIMEUNIT', target=target_col) 	
df_noWheels.calc_frequency(col='Model', target=target_col)
df_noWheels.calc_frequency(col='Nationality', target=target_col)
df_noWheels.calc_frequency(col='Size', target=target_col)
df_noWheels.calc_frequency(col='Trim', target=target_col) 	
df_noWheels.calc_frequency(col='SubModel', target=target_col)
### can be important
df_noWheels.calc_frequency(col='VNST', target=target_col).head(10)
df_noWheels.calc_frequency(col='VehicleAge', target=target_col).sort_values('IsBadBuy_%') 
df_noWheels.calc_frequency(col='BYRNO', target=target_col).head(10)


df_noWheels.calc_frequency(col='Make', target=target_col).head(20)

df_noWheels.calc_frequency(col='VNZIP1', target=target_col).head(10)

### Conclusion for no Wheels data, important columns:
'Auction', 'VNZIP1', 'BYRNO', 'VehicleAge'

df_Wheels.cols_importance(cols=cols, target='IsBadBuy').sort_values('delta1000', ascending=False)

df_Wheels.calc_frequency(col='VNZIP1', target=target_col).head(10)

### Conclusion 
coloumns to drop: Nationality, IsOnlineSale, Transmission, Color, TopThreeAmericanName, PRIMEUNIT, AUCGUART, Size, VNST, VehYear
