==> notebooks_clean/03_EDA_numerical.ipynb <==
import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns
from avp_pckg.DataFrame import AvPdataFrame 

%load_ext autoreload
%autoreload 2

df = pd.read_csv('data\\df_train.csv', parse_dates=['PurchDate'], index_col=0)
df.columns
cols = ['IsBadBuy','VehicleAge', 'Make', 'Model', 'WheelType', 'VehOdo',
        'MMRAcquisitionAuctionAveragePrice', 'MMRAcquisitionAuctionCleanPrice',
       'MMRAcquisitionRetailAveragePrice', 'MMRAcquisitonRetailCleanPrice',
       'MMRCurrentAuctionAveragePrice', 'MMRCurrentAuctionCleanPrice',
       'MMRCurrentRetailAveragePrice', 'MMRCurrentRetailCleanPrice',
       'VehBCost', 'WarrantyCost',
       ]
df = df.loc[:, cols].copy()
df.loc[ :, 'WheelType'] = df['WheelType'].fillna(value='empty')
df.fillna(value=0, inplace=True)
df = AvPdataFrame(df)

### Price: Average - Clean, Auction - Retail, Acquisition - Current

df.info()

# 'MMRAcquisitionAuctionAveragePrice', 'MMRAcquisitionAuctionCleanPrice'
df.loc[:, 'AcqAuc'] = df['MMRAcquisitionAuctionCleanPrice'] - df['MMRAcquisitionAuctionAveragePrice'] # Clean - Average
# 'MMRAcquisitionRetailAveragePrice', 'MMRAcquisitonRetailCleanPrice'
df.loc[:, 'AcqRetail'] = df['MMRAcquisitonRetailCleanPrice'] - df['MMRAcquisitionRetailAveragePrice'] # Clean - Average

df.loc[:, 'AcqClean'] = df['MMRAcquisitonRetailCleanPrice'] - df['MMRAcquisitionAuctionCleanPrice'] # Retail - Auction
df.loc[:, 'RetailClean'] = df['MMRAcquisitonRetailCleanPrice'] - df['MMRCurrentRetailCleanPrice'] # Acquisiton - Current

fig, axs = plt.subplots(nrows=2, ncols=2, figsize=(10, 10))
sns.scatterplot(df, x='MMRAcquisitionAuctionCleanPrice', y='AcqAuc', hue='IsBadBuy', s=5, ax=axs[0][0]) # Clean - Average
sns.scatterplot(df, x='MMRAcquisitonRetailCleanPrice', y='AcqRetail', hue='IsBadBuy', s=5, ax=axs[1][0]) # Clean - Average
sns.scatterplot(df, x='MMRAcquisitonRetailCleanPrice', y='AcqClean', hue='IsBadBuy',s=5, ax=axs[0][1]) # Retail - Auction
sns.scatterplot(df, x='MMRAcquisitonRetailCleanPrice', y='RetailClean', hue='IsBadBuy',s=5, ax=axs[1][1]) # Acquisiton - Current

df.columns
df.calc_frequency(col='Make', target='IsBadBuy').head(5)

df_CHEVROLET = df[df['Make'] == 'CHEVROLET']

fig, axs = plt.subplots(nrows=2, ncols=2, figsize=(10, 10))
sns.scatterplot(df_CHEVROLET, x='MMRAcquisitionAuctionCleanPrice', y='AcqAuc', hue='IsBadBuy', s=5, ax=axs[0][0]) ## Clean - Average  'VehicleAge' 'IsBadBuy' 'MMRAcquisitionAuctionAveragePrice'
sns.scatterplot(df_CHEVROLET, x='MMRAcquisitonRetailCleanPrice', y='AcqRetail', hue='IsBadBuy', s=5, ax=axs[1][0]) # Clean - Average

sns.scatterplot(df_CHEVROLET, x='MMRAcquisitonRetailCleanPrice', y='AcqClean', hue='IsBadBuy', s=5, ax=axs[0][1]) # Retail - Auction
sns.scatterplot(df_CHEVROLET, x='MMRAcquisitonRetailCleanPrice', y='RetailClean', hue='IsBadBuy', s=5, ax=axs[1][1]) # Acquisiton - Current

df.columns
df.calc_frequency(col='Model', target='IsBadBuy').head(5)

df_PT_CRUISER = df[df['Model'] == 'PT CRUISER']

fig, axs = plt.subplots(nrows=2, ncols=2, figsize=(10, 10))
sns.scatterplot(df_PT_CRUISER, x='MMRAcquisitionAuctionCleanPrice', y='AcqAuc', hue='IsBadBuy',s=5, ax=axs[0][0]) ## Clean - Average  'VehicleAge' 'IsBadBuy' 'MMRAcquisitionAuctionAveragePrice'
sns.scatterplot(df_PT_CRUISER, x='MMRAcquisitonRetailCleanPrice', y='AcqRetail', hue='IsBadBuy',s=5, ax=axs[1][0]) # Clean - Average

sns.scatterplot(df_PT_CRUISER, x='MMRAcquisitonRetailCleanPrice', y='AcqClean', hue='IsBadBuy',s=5, ax=axs[0][1]) # Retail - Auction
sns.scatterplot(df_PT_CRUISER, x='MMRAcquisitonRetailCleanPrice', y='RetailClean', hue='IsBadBuy',s=5, ax=axs[1][1]) # Acquisiton - Current


### Split: with and without wheels

df_noWheels = df[df['WheelType'] == 'empty']
df_Wheels = df[df['WheelType'] != 'empty']
df_noWheels.head()

fig, axs = plt.subplots(nrows=2, ncols=2, figsize=(10, 10))
sns.scatterplot(df_noWheels, x='MMRAcquisitionAuctionCleanPrice', y='AcqAuc', hue='IsBadBuy', s=5, ax=axs[0][0]) # Clean - Average
sns.scatterplot(df_noWheels, x='MMRAcquisitonRetailCleanPrice', y='AcqRetail', hue='IsBadBuy', s=5, ax=axs[1][0]) # Clean - Average
sns.scatterplot(df_noWheels, x='MMRAcquisitonRetailCleanPrice', y='AcqClean', hue='IsBadBuy',s=5, ax=axs[0][1]) # Retail - Auction
sns.scatterplot(df_noWheels, x='MMRAcquisitonRetailCleanPrice', y='RetailClean', hue='IsBadBuy',s=5, ax=axs[1][1]) # Acquisiton - Current


fig, axs = plt.subplots(nrows=2, ncols=2, figsize=(10, 10))
sns.scatterplot(df_Wheels, x='MMRAcquisitionAuctionCleanPrice', y='AcqAuc', hue='IsBadBuy', s=5, ax=axs[0][0]) # Clean - Average
sns.scatterplot(df_Wheels, x='MMRAcquisitonRetailCleanPrice', y='AcqRetail', hue='IsBadBuy', s=5, ax=axs[1][0]) # Clean - Average
sns.scatterplot(df_Wheels, x='MMRAcquisitonRetailCleanPrice', y='AcqClean', hue='IsBadBuy',s=5, ax=axs[0][1]) # Retail - Auction
sns.scatterplot(df_Wheels, x='MMRAcquisitonRetailCleanPrice', y='RetailClean', hue='IsBadBuy',s=5, ax=axs[1][1]) # Acquisiton - Current

df_noWheels = AvPdataFrame(df_noWheels)
df_noWheels.calc_frequency(col='Model', target='IsBadBuy').head(5)

df_PT_CRUISER = df_noWheels[df_noWheels['Model'] == 'PT CRUISER']

fig, axs = plt.subplots(nrows=2, ncols=2, figsize=(10, 10))
sns.scatterplot(df_PT_CRUISER, x='MMRAcquisitionAuctionCleanPrice', y='AcqAuc', hue='IsBadBuy',s=5, ax=axs[0][0]) ## Clean - Average  'VehicleAge' 'IsBadBuy' 'MMRAcquisitionAuctionAveragePrice'
sns.scatterplot(df_PT_CRUISER, x='MMRAcquisitonRetailCleanPrice', y='AcqRetail', hue='IsBadBuy',s=5, ax=axs[1][0]) # Clean - Average

sns.scatterplot(df_PT_CRUISER, x='MMRAcquisitonRetailCleanPrice', y='AcqClean', hue='IsBadBuy',s=5, ax=axs[0][1]) # Retail - Auction
sns.scatterplot(df_PT_CRUISER, x='MMRAcquisitonRetailCleanPrice', y='RetailClean', hue='IsBadBuy',s=5, ax=axs[1][1]) # Acquisiton - Current

df.columns

cols_num =['IsBadBuy', 'VehicleAge', 'VehOdo',
       'MMRAcquisitionAuctionAveragePrice', 'MMRAcquisitionAuctionCleanPrice',
       'MMRAcquisitionRetailAveragePrice', 'MMRAcquisitonRetailCleanPrice',
       'MMRCurrentAuctionAveragePrice', 'MMRCurrentAuctionCleanPrice',
       'MMRCurrentRetailAveragePrice', 'MMRCurrentRetailCleanPrice',
       'VehBCost', 'WarrantyCost', 'AcqAuc', 'AcqRetail', 'AcqClean',
       'RetailClean']
df_num = df.loc[:, cols_num]

plt.figure(figsize=(15, 10))
cor = df_num.corr()
# sns.heatmap(cor, annot=True, cmap=plt.cm.Reds)
sns.heatmap(cor, annot=True, cmap='coolwarm', vmin = -1, vmax = 1, fmt='.2f')

df.columns

df_noWheels = AvPdataFrame(df_noWheels)
df_Wheels = AvPdataFrame(df_Wheels)
df_cost = df.calc_frequency_num(col='VehBCost', target='IsBadBuy')
df_cost_nW = df_noWheels.calc_frequency_num(col='VehBCost', target='IsBadBuy')
df_cost_Wh = df_Wheels.calc_frequency_num(col='VehBCost', target='IsBadBuy')

df_cost_nW.columns


df_Wheels['IsBadBuy'].sum()/ df_Wheels.shape[0]

df_cost_nW
AvPdataFrame.plot_frequency_num(df=df_cost,
                        x_vlines=[3500, 7000, 12000], 
                        zoomX1=[800, 6000], 
                        zoomX2=[5000, 22000], 
                        xHist=[800, 15000],
                        bins=100
                        )

AvPdataFrame.plot_frequency_num(df=df_cost_Wh,
                        x_vlines=[3500, 7000, 12000], 
                        zoomX1=[800, 6000], 
                        zoomX2=[5000, 22000], 
                        xHist=[800, 15000],
                        bins=100
                        )

AvPdataFrame.plot_frequency_num(df=df_cost_nW,
                        x_vlines=[3500, 7000, 12000], 
                        zoomX1=[800, 6000], 
                        zoomX2=[5000, 22000], 
                        xHist=[800, 15000],
                        bins=100
                        )

df_noWheels = AvPdataFrame(df_noWheels)
df_Wheels = AvPdataFrame(df_Wheels)
df_VehOdof = df.calc_frequency_num(col='VehOdo', target='IsBadBuy')
df_VehOdof_nW = df_noWheels.calc_frequency_num(col='VehOdo', target='IsBadBuy')
df_VehOdof_Wh = df_Wheels.calc_frequency_num(col='VehOdo', target='IsBadBuy')


df_VehOdof = df.calc_frequency_num(col='VehOdo', target='IsBadBuy', bin=100)
AvPdataFrame.plot_frequency_num(df=df_VehOdof,
                        x_vlines=[50_000, 75_000, 100_000], 
                        lines_color='r',
                        zoomX1=[5_000, 60_000], 
                        zoomX2=[50_000, 120_000], 
                        xHist=None,
                       # xHist=[10_000, 120_000]
                        hist_bins=100
                        )

df_VehOdof_nW = df_noWheels.calc_frequency_num(col='VehOdo', target='IsBadBuy', bin=100)
AvPdataFrame.plot_frequency_num(df=df_VehOdof_nW,
                        x_vlines=[53_000, 70_000, 79_000, 83_000], 
                        lines_color='g',
                        zoomX1=[43_000, 80_000], 
                        zoomX2=[70_000, 110_000], 
                        xHist=None,
                       # xHist=[10_000, 120_000],
                        hist_bins=100
                        )

df_VehOdof_Wh = df_Wheels.calc_frequency_num(col='VehOdo', target='IsBadBuy', bin=1000)
AvPdataFrame.plot_frequency_num(df=df_VehOdof_Wh,
                        x_vlines=[50_000, 75_000, 100_000], 
                        lines_color='r',
                        zoomX1=[5_000, 60_000], 
                        zoomX2=[50_000, 120_000], 
                        xHist=None,
                       # xHist=[10_000, 120_000]
                        hist_bins=100
                        )

# Diferenses

## AcqAuc

```python 
df.loc[:, 'AcqAuc'] = df['MMRAcquisitionAuctionCleanPrice'] - df['MMRAcquisitionAuctionAveragePrice'] #
``` 
Clean - Average

- full data

df_AcqAuc = df.calc_frequency_num(col='AcqAuc', target='IsBadBuy', bin=500)

# 'AcqAuc', 'AcqRetail', 'AcqClean', 'RetailClean'

AvPdataFrame.plot_frequency_num(df=df_AcqAuc,
                        x_vlines=[400, 600, 1_100], 
                        lines_color='k',
                        zoomX1=[0, 1_000], 
                        zoomX2=[800, 3_000], 
                        xHist=None,
                       # xHist=[10_000, 120_000]
                        hist_bins=100
                        )

- with Wheels

df_AcqAu_W = df_Wheels.calc_frequency_num(col='AcqAuc', target='IsBadBuy', bin=500)

AvPdataFrame.plot_frequency_num(df=df_AcqAu_W,
                        x_vlines=[400, 600, 1_100], 
                        lines_color='darkorange',
                        zoomX1=[0, 1_000], 
                        zoomX2=[800, 3_000], 
                        xHist=None,
                       # xHist=[10_000, 120_000]
                        hist_bins=100
                        )

- no Wheels

df_AcqAu_nW = df_noWheels.calc_frequency_num(col='AcqAuc', target='IsBadBuy', bin=300)

AvPdataFrame.plot_frequency_num(df=df_AcqAu_nW,
                        x_vlines=[ 1_000], 
                        lines_color='k',
                        zoomX1=[0, 1_000], 
                        zoomX2=[800, 3_000], 
                        xHist=None,
                       # xHist=[10_000, 120_000]
                        hist_bins=100
                        )

## AcqRetail 

```python
df.loc[:, 'AcqRetail'] = df['MMRAcquisitonRetailCleanPrice'] - df['MMRAcquisitionRetailAveragePrice'] 
```
Clean - Average

- full data 

# 'AcqAuc', 'AcqRetail', 'AcqClean', 'RetailClean'
df_AcqRetail = df.calc_frequency_num(col='AcqRetail', target='IsBadBuy', bin=500)


AvPdataFrame.plot_frequency_num(df_AcqRetail,
                        x_vlines=[200, 400, 650, 1_200], 
                        lines_color='k',
                        zoomX1=[0, 1_000], 
                        zoomX2=[800, 3_000], 
                        xHist=None,
                       # xHist=[10_000, 120_000]
                        hist_bins=100
                        )

- with Wheels

# 'AcqAuc', 'AcqRetail', 'AcqClean', 'RetailClean'
df_AcqRetail_W = df_Wheels.calc_frequency_num(col='AcqRetail', target='IsBadBuy', bin=500)

AvPdataFrame.plot_frequency_num(df_AcqRetail_W,
                        x_vlines=[200, 650, 1_200], 
                        lines_color='darkorange',
                        zoomX1=[0, 1_000], 
                        zoomX2=[800, 3_000], 
                        xHist=None,
                       # xHist=[10_000, 120_000]
                        hist_bins=100
                        )

- no Wheels

df_AcqRetail_nW = df_noWheels.calc_frequency_num(col='AcqRetail', target='IsBadBuy', bin=100)

AvPdataFrame.plot_frequency_num(df_AcqRetail_nW,
                        x_vlines=[200, 650, 1_200], 
                        lines_color='k',
                        zoomX1=[0, 1_000], 
                        zoomX2=[800, 3_000], 
                        xHist=None,
                       # xHist=[10_000, 120_000]
                        hist_bins=100
                        )

## AcqClean
```python
df.loc[:, 'AcqClean'] = df['MMRAcquisitonRetailCleanPrice'] - df['MMRAcquisitionAuctionCleanPrice'] 
```
 Retail - Auction

- full data

# 'AcqAuc', 'AcqRetail', 'AcqClean', 'RetailClean'
df_AcqClean = df.calc_frequency_num(col='AcqClean', target='IsBadBuy', bin=300)


AvPdataFrame.plot_frequency_num(df=df_AcqClean,
                        x_vlines=[100, 550, 720, 1000, 2000, 3000], 
                        lines_color='k',
                        zoomX1=[0, 1_500], 
                        zoomX2=[800, 6_000], 
                        xHist=None,
                       # xHist=[10_000, 120_000]
                        hist_bins=100
                        )

- with Wheels

df_Wheels = AvPdataFrame(df_Wheels)
df_AcqClean_W = df_Wheels.calc_frequency_num(col='AcqClean', target='IsBadBuy', bin=300,)

AvPdataFrame.plot_frequency_num(df=df_AcqClean_W,
                        x_vlines=[300, 800, 1000, 2000, 3000], 
                        lines_color='darkorange',
                        zoomX1=[0, 1_500], 
                        zoomX2=[800, 6_000], 
                        xHist=None,
                       # xHist=[10_000, 120_000]
                        hist_bins=100
                        )

- no Wheels

df_noWheels = AvPdataFrame(df_noWheels)
df_AcqClean_nW = df_noWheels.calc_frequency_num(col='AcqClean', target='IsBadBuy', bin=100,)


AvPdataFrame.plot_frequency_num(df=df_AcqClean_nW,
                        x_vlines=[1000, 2000, 3700], 
                        lines_color='r',
                        zoomX1=[400, 1_500], 
                        zoomX2=[800, 6_000], 
                        xHist=None,
                       # xHist=[10_000, 120_000]
                        hist_bins=100
                        )

## 'RetailClean'
```python
df.loc[:, 'RetailClean'] = df['MMRAcquisitonRetailCleanPrice'] - df['MMRCurrentRetailCleanPrice'] # 
```
Acquisiton - Current

- full data

df_RetailClean = df.calc_frequency_num(col='RetailClean', target='IsBadBuy', bin=300,)

# 'AcqAuc', 'AcqRetail', 'AcqClean', 'RetailClean'

AvPdataFrame.plot_frequency_num(df=df_RetailClean,
                        x_vlines=[-2_000, -1200, -450, -100, 100, 450, ], 
                        lines_color='k',
                        zoomX1=[-3_000, 0], 
                        zoomX2=[-1_000, 2_000], 
                        xHist=None,
                       # xHist=[10_000, 120_000]
                        hist_bins=100
                        )

- with Wheels

df_Wheels = AvPdataFrame(df_Wheels)
df_RetailClean_W = df_Wheels.calc_frequency_num(col='RetailClean', target='IsBadBuy', bin=1000,)

AvPdataFrame.plot_frequency_num(df=df_RetailClean_W ,
                        x_vlines=[-5_000, -1_200, -150, 450, ], 
                        lines_color='darkorange',
                        zoomX1=[-10_000, 450], 
                        zoomX2=[-2_000, 2_000], 
                        xHist=None,
                       # xHist=[10_000, 120_000]
                        hist_bins=100
                        )

- no Wheels

df_noWheels = AvPdataFrame(df_noWheels)
df_RetailClean_nW = df_noWheels.calc_frequency_num(col='RetailClean', target='IsBadBuy', bin=50,)


AvPdataFrame.plot_frequency_num(df=df_RetailClean_nW ,
                        x_vlines=[-2_500, -1200, -450, -100, 100, 450, ], 
                        lines_color='r',
                        zoomX1=[-3_000, 0], 
                        zoomX2=[-1_000, 2_000], 
                        xHist=None,
                       # xHist=[10_000, 120_000]
                        hist_bins=10
                        )

sns.scatterplot(df_noWheels,
                x='RetailClean',
                y='IsBadBuy')
# df_noWheels['RetailClean']

df_tmp = df_noWheels[['IsBadBuy','RetailClean']].copy().sort_values('RetailClean', ignore_index=True, ascending=True)
df_tmp.reset_index(inplace=True)
df_tmp.head()




target_freq = df_tmp.loc[0:4, 'IsBadBuy'].sum() # / 1
target_freq


df.loc[:, 'RetailClean'] = df['MMRAcquisitonRetailCleanPrice'] - df['MMRCurrentRetailCleanPrice']
### no Wheels: x_vlines=[-2_000, -1200, -450, -100, 100, 450, ]
### Wheels:    x_vlines=[-5_000, -1200, -150, 450, ]

df.loc[:, 'AcqClean'] = df['MMRAcquisitonRetailCleanPrice'] - df['MMRAcquisitionAuctionCleanPrice']
### no Wheels:          x_vlines=[1000, 2000, 3700]
### Wheels:   x_vlines=[300, 800, 1000, 2000, 3000] 

df.loc[:, 'AcqRetail'] = df['MMRAcquisitonRetailCleanPrice'] - df['MMRAcquisitionRetailAveragePrice'] 
### Wheels: x_vlines=[200, 650, 1_200]

df.loc[:, 'AcqAuc'] = df['MMRAcquisitionAuctionCleanPrice'] - df['MMRAcquisitionAuctionAveragePrice']
### Wheels: x_vlines=[400, 600, 1_100]
